Restructure packaging tests to make format-specific test modules.
