+++ tests/platforms/macOS/app/package/conftest.py
@@ -1,0 +1,26 @@
+import subprocess
+from unittest import mock
+
+import pytest
+
+from briefcase.console import Console, Log
+from briefcase.platforms.macOS.app import macOSAppPackageCommand
+
+
+@pytest.fixture
+def package_command(tmp_path):
+    command = macOSAppPackageCommand(
+        logger=Log(),
+        console=Console(),
+        base_path=tmp_path / "base_path",
+        data_path=tmp_path / "briefcase",
+    )
+
+    command.select_identity = mock.MagicMock()
+    command.sign_app = mock.MagicMock()
+    command.sign_file = mock.MagicMock()
+    command.notarize = mock.MagicMock()
+    command.dmgbuild = mock.MagicMock()
+    command.tools.subprocess = mock.MagicMock(spec=subprocess)
+
+    return command
