+++ tests/platforms/macOS/app/package/test_notarize.py
